requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==1.4
psycopg2-binary
boto3
docxtpl
openpyxl
pandas
python-dateutil
phonenumbers
pdfplumber
python-docx
pytesseract
openai
jinja2
jinja2-cli
pytest
tenacity
rapidfuzz
python-multipart
python-magic

